add loss weighting

+++ training_utils_atac_rna.py
@@ -437,7 +437,7 @@ def deserialize_tr(serialized_example, g, use_tf_activity, input_length = 196608
                 tf.cast(tf.ensure_shape(atac_out,[output_length-crop_size*2,1]),dtype=tf.float32), \
                 tf.cast(tf.ensure_shape(rna_out,[output_length-crop_size*2,1]),dtype=tf.float32), \
                 tf.cast(tf.ensure_shape(rna_assay_type,[]),dtype=tf.int32), \
-                tf.cast(tf.ensure_shape(weighting_factor,[]),dtype=tf.int32), \
+                tf.cast(tf.ensure_shape(weighting_factor,[]),dtype=tf.float32), \
                 tf.cast(tf.ensure_shape(tf_activity, [1,1629]),dtype=tf.bfloat16)
 
 def deserialize_val(serialized_example, g, use_tf_activity, input_length = 196608,
@@ -610,7 +610,7 @@ def deserialize_val(serialized_example, g, use_tf_activity, input_length = 19660
                 tf.cast(tf.ensure_shape(atac_out,[output_length-crop_size*2,1]),dtype=tf.float32), \
                 tf.cast(tf.ensure_shape(rna_out,[output_length-crop_size*2,1]),dtype=tf.float32), \
                 tf.cast(tf.ensure_shape(rna_assay_type,[]),dtype=tf.int32), \
-                tf.cast(tf.ensure_shape(weighting_factor,[]),dtype=tf.int32), \
+                tf.cast(tf.ensure_shape(weighting_factor,[]),dtype=tf.float32), \
                 tf.cast(tf.ensure_shape(tf_activity, [1,1629]),dtype=tf.bfloat16)
 
 def deserialize_val_TSS(serialized_example, g, use_tf_activity, input_length = 196608,
